Fix captions
This fixes captions not being able to have links in them. Previously links in captions where stripped and additionally not clickable as the JS for the click event controlled the video playback was overiding the click.

This also enables support for all HTML allowed in the caption including bold, italic, inline images, code, strikethrough, superscript, sub script and keyboard input.